Fix: organizes files and fixes inputs
